requirements.txt
# χ²-DFD 深度伪造检测系统 - 依赖包列表

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 自然语言处理和多模态模型
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.39.0

# 图像处理
Pillow>=9.0.0
opencv-python-headless>=4.8.0

# 数值计算
numpy>=1.21.0

# 网络请求
requests>=2.28.0

# 文件格式支持
safetensors>=0.3.0

# 模型加载和缓存
huggingface-hub>=0.15.0
filelock>=3.12.0

# 其他工具
psutil>=5.9.0
